requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
class-registry==2.1.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django~=5.0.6
django-filter==24.2
django-graphql-jwt==0.4.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
filters==1.3.2
filters-django==1.0.5
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
promise==2.3
prompt_toolkit==3.0.46
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
uritemplate==4.1.1
wcwidth==0.2.13
pillow~=10.4.0